fix: move zscores due to circular import
