+++ src/reversion/optimize_reversion.py
@@ -2,8 +2,8 @@ from typing import Dict, List, Tuple
 import optuna
 import pandas as pd
 
-from reversion.multiscale_reversion import calculate_robust_z_scores
-from reversion.reversion_utils import composite_score, simulate_strategy
+from reversion.strategy_metrics import composite_score, simulate_strategy
+from reversion.z_scores import calculate_robust_z_scores
 from utils.logger import logger
 from utils.caching_utils import load_parameters_from_pickle, save_parameters_to_pickle
