+++ src/reversion/strategy_metrics.py
@@ -1,4 +1,5 @@
 from typing import Any, Dict, Tuple
+import numpy as np
 import pandas as pd
 
 from utils.performance_metrics import kappa_ratio, sharpe_ratio
