+++ src/reversion/z_scores.py
@@ -4,6 +4,22 @@ import plotly.graph_objects as go
 from plotly.subplots import make_subplots
 
 
+def calculate_robust_z_scores(returns_df: pd.DataFrame, window: int) -> pd.DataFrame:
+    """
+    Calculate rolling robust z-scores using the rolling median and MAD.
+    The MAD is scaled by 1.4826 to approximate the standard deviation.
+    """
+    rolling_median = returns_df.rolling(window=window, min_periods=1).median()
+    # Rolling MAD: use a lambda to compute median absolute deviation over the window
+    mad = returns_df.rolling(window=window, min_periods=1).apply(
+        lambda x: np.median(np.abs(x - np.median(x))), raw=True
+    )
+    # Avoid division by zero by replacing zeros in mad
+    mad.replace(0, np.nan, inplace=True)
+    robust_z = (returns_df - rolling_median) / (mad * 1.4826)
+    return robust_z.fillna(0)
+
+
 def plot_robust_z_scores(
     robust_z: pd.DataFrame,
     z_threshold: float,
